+++ $Window.js
@@ -9,9 +9,28 @@ function find_tabstops($el) {
 	// This function finds focusable controls, but not necessarily all of them;
 	// for radio elements, it only gives one: either the checked one, or the first one if none are checked.
 
-	let $controls = $el.find("input, textarea, select, button, object, a[href], [tabIndex='0'], details summary")
-		.filter(":enabled, summary, a")
-		.filter(":visible");
+	// Note: for audio[controls], Chrome at least has two tabstops (the audio element and three dots menu button).
+	// It might be possible to detect this in the shadow DOM, I don't know, I haven't worked with the shadow DOM.
+	// But it might be more reliable to make a dummy tabstop element to detect when you tab out of the first/last element.
+	// Assuming that doesn't mess with screen readers.
+	// Right now you can't tab to the three dots menu if it's the last element.
+	// @TODO: see what ally.js does. Does it handle audio[controls]? https://allyjs.io/api/query/tabsequence.html
+
+	let $controls = $el.find(`
+		input:enabled,
+		textarea:enabled,
+		select:enabled,
+		button:enabled,
+		a[href],
+		[tabIndex='0'],
+		details summary,
+		iframe,
+		object,
+		embed,
+		video[controls],
+		audio[controls],
+		[contenteditable]:not([contenteditable='false'])
+	`).filter(":visible");
 	// const $controls = $el.find(":tabbable"); // https://api.jqueryui.com/tabbable-selector/
 
 	// Radio buttons should be treated as a group with one tabstop.